Separate the generation of the Checkpoint from the CheckpointSchedule
In previous commits the CheckpointSchedule controller would be
responsible for creating the CronJob to run the CheckpointSchedule
where it would also create the Checkpoint at every execution of the
CronJob.

Now, we introduced a CheckpointRequest resource that is created in
the CheckpointSchedule CronJob. Then, at the CheckpointRequest
controller we actually create a new Checkpoint. This decouples the
Checkpoint logic from the schedule and allow better testability and
scalability.